feat(story): delete first page on story delete

+++ sql/schema.sql
@@ -10,5 +10,5 @@ CREATE TABLE story
   id BIGSERIAL PRIMARY KEY,
   title VARCHAR(32) NOT NULL,
   description VARCHAR(512),
-  first_page_id int REFERENCES page(id)
+  first_page_id BIGSERIAL REFERENCES page(id)
 );
